feat: Refactor logo handling to use Uint8List for logo URLs and replace NetworkImageWidget for better image loading
